requirements.txt
TeslaPy==2.8.0
snowflake-connector-python==3.10.1
python-dotenv==1.0.1
